Move presentation logic to presentation manager
