+++ android/src/main/java/com/swmansion/rnscreens/gamma/modals/formsheet/FormSheetDialogManager.kt
@@ -1,6 +1,5 @@
 package com.swmansion.rnscreens.gamma.modals.formsheet
 
-import android.animation.Animator
 import android.content.Context
 import android.util.Log
 import android.view.ContextThemeWrapper
@@ -41,7 +40,7 @@ class FormSheetDialogManager(
 
     private val dimmingManager = DimmingViewManager(context, dialog)
 
-    private val animationCoordinator = FormSheetAnimatorFactory(dimmingManager)
+    private val animatorFactory = FormSheetAnimatorFactory(dimmingManager)
 
     private val dimensionsCoordinator =
         FormSheetDimensionsCoordinator(
@@ -51,63 +50,12 @@ class FormSheetDialogManager(
             behaviorController = behaviorController,
         )
 
-    private var currentSheetAnimator: Animator? = null
-
     private val presentationManager =
         FormSheetPresentationManager(
-            performPresent = { onComplete ->
-                dialog.setOnShowListener {
-                    dialog.setOnShowListener(null)
-                    dimmingManager.onDialogShow()
-
-                    if (bottomSheetView == null) {
-                        onComplete()
-                        return@setOnShowListener
-                    }
-
-                    val isInterrupting = currentSheetAnimator?.isRunning == true
-                    currentSheetAnimator?.cancel()
-
-                    currentSheetAnimator = animationCoordinator.createEnterAnimator(bottomSheetView, isInterrupting).apply {
-                        addListener(object : android.animation.AnimatorListenerAdapter() {
-                            override fun onAnimationEnd(animation: Animator) {
-                                if (currentSheetAnimator == this@apply) currentSheetAnimator = null
-                                onComplete()
-                            }
-                        })
-                        start()
-                    }
-                }
-                dialog.show()
-            },
-            performDismiss = { onComplete ->
-                if (bottomSheetView == null) {
-                    dialog.dismiss()
-                    onComplete()
-                    return@FormSheetPresentationManager
-                }
-
-                val behavior = BottomSheetBehavior.from(bottomSheetView)
-                if (behavior.state == BottomSheetBehavior.STATE_HIDDEN) {
-                    dialog.dismiss()
-                    onComplete()
-                    return@FormSheetPresentationManager
-                }
-
-                val isInterrupting = currentSheetAnimator?.isRunning == true
-                currentSheetAnimator?.cancel()
-
-                currentSheetAnimator = animationCoordinator.createExitAnimator(bottomSheetView, isInterrupting).apply {
-                    addListener(object : android.animation.AnimatorListenerAdapter() {
-                        override fun onAnimationEnd(animation: Animator) {
-                            if (currentSheetAnimator == this@apply) currentSheetAnimator = null
-                            dialog.dismiss()
-                            onComplete()
-                        }
-                    })
-                    start()
-                }
-            },
+            dialog = dialog,
+            bottomSheetView = bottomSheetView,
+            dimmingManager = dimmingManager,
+            animatorFactory = animatorFactory,
             onNativeDismiss = onDismissRequest,
         )
